Hidden Markov Models

A Hidden Markov Model (HMM) is a time series model with unobservable latent variables that follow a Markov process. Condititional distributions (when conditioned on the latents) are often chosen to be common distributions (e.g. Gaussian).

StateSpaceDynamics.HiddenMarkovModel — Type

HiddenMarkovModel

A Hidden Markov Model (HMM) with custom emissions.

Fields

K::Int: Number of states.

B::Vector=Vector(): Vector of emission models.

emission=nothing: If B is missing emissions, clones of this model will be used to fill in the rest.

A::Matrix{<:Real}: Transition matrix.

πₖ::Vector{Float64}: Initial state distribution.

StateSpaceDynamics.class_probabilities — Function

function class_probabilities(model::HiddenMarkovModel, Y::Matrix{<:Real}, X::Union{Matrix{<:Real},Nothing}=nothing;)

Calculate the class probabilities at each time point using forward backward algorithm

Arguments

model::HiddenMarkovModel: The Hidden Markov Model to fit.

Y::Matrix{<:Real}: The emission data

X::Union{Matrix{<:Real},Nothing}=nothing: Optional input data for fitting Switching Regression Models

Returns

class_probabilities::Matrix{Float64}: The class probabilities at each timepoint

StateSpaceDynamics.class_probabilities — Function

function class_probabilities(model::HiddenMarkovModel, Y::Vector{<:Matrix{<:Real}}, X::Union{Vector{<:Matrix{<:Real}},Nothing}=nothing;)

Calculate the class probabilities at each time point using forward backward algorithm on multiple trials of data

Arguments

model::HiddenMarkovModel: The Hidden Markov Model to fit.

Y::Vectpr{<:Matrix{<:Real}}: The trials of emission data

X::Union{Vector{<:Matrix{<:Real}},Nothing}=nothing: Optional trials of input data for fitting Switching Regression Models

Returns

class_probabilities::Vector{<:Matrix{Float64}}: Each trial's class probabilities at each timepoint

StateSpaceDynamics.fit! — Function

fit!(model::HiddenMarkovModel, Y::Matrix{<:Real}, X::Union{Matrix{<:Real}, Nothing}=nothing; max_iters::Int=100, tol::Float64=1e-6)

Fit the Hidden Markov Model using the EM algorithm.

Arguments

model::HiddenMarkovModel: The Hidden Markov Model to fit.

Y::Matrix{<:Real}: The emission data.

X::Union{Matrix{<:Real}, Nothing}=nothing: Optional input data for fitting Switching Regression Models

max_iters::Int=100: The maximum number of iterations to run the EM algorithm.

tol::Float64=1e-6: When the log likelihood is improving by less than this value, the algorithm will stop.

StateSpaceDynamics.fit! — Function

fit!(model::HiddenMarkovModel, Y::Matrix{<:Real}, X::Union{Matrix{<:Real}, Nothing}=nothing; max_iters::Int=100, tol::Float64=1e-6)

Fit the Hidden Markov Model to multiple trials of data using the EM algorithm.

Arguments

model::HiddenMarkovModel: The Hidden Markov Model to fit.

Y::Vector{<:Matrix{<:Real}}: The trialized emission data.

X::Union{Vector{<:Matrix{<:Real}}, Nothing}=nothing: Optional input data for fitting Switching Regression Models

max_iters::Int=100: The maximum number of iterations to run the EM algorithm.

tol::Float64=1e-6: When the log likelihood is improving by less than this value, the algorithm will stop.

StateSpaceDynamics.sample — Function

sample(model::HiddenMarkovModel, data...; n::Int)

Generate n samples from a Hidden Markov Model. Returns a tuple of the state sequence and the observation sequence.

Arguments

model::HiddenMarkovModel: The Hidden Markov Model to sample from.

data...: The data to fit the Hidden Markov Model. Requires the same format as the emission model.

n::Int: The number of samples to generate.

Returns

state_sequence::Vector{Int}: The state sequence, where each element is an integer 1:K.

observation_sequence::Matrix{Float64}: The observation sequence. This takes the form of the emission model's output.

StateSpaceDynamics.viterbi — Function

viterbi(model::HiddenMarkovModel, Y::Matrix{<:Real}, X::Union{Matrix{<:Real},Nothing}=nothing;)

Get most likely class labels using the Viterbi algorithm

Arguments

model::HiddenMarkovModel: The Hidden Markov Model to fit.

Y::Matrix{<:Real}: The emission data

X::Union{Matrix{<:Real},Nothing}=nothing: Optional input data for fitting Switching Regression Models

Returns

best_path::Vector{Float64}: The most likely state label at each timepoint

StateSpaceDynamics.viterbi — Function

viterbi(model::HiddenMarkovModel, Y::Vector{<:Matrix{<:Real}}, X::Union{Vector{<:Matrix{<:Real}},Nothing}=nothing;)

Get most likely class labels using the Viterbi algorithm for multiple trials of data

Arguments

model::HiddenMarkovModel: The Hidden Markov Model to fit.

Y::Vectpr{<:Matrix{<:Real}}: The trials of emission data

X::Union{Vector{<:Matrix{<:Real}},Nothing}=nothing: Optional trials of input data for fitting Switching Regression Models

Returns

best_path::Vector{<:Vector{Float64}}: Each trial's best state path

StateSpaceDynamics.GaussianHMM — Method

GaussianHMM(; K::Int, output_dim::Int, A::Matrix{<:Real}=initialize_transition_matrix(K), πₖ::Vector{Float64}=initialize_state_distribution(K))

Create a Hidden Markov Model with Gaussian Emissions

Arguments

K::Int: The number of hidden states

output_dim::Int: The dimensionality of the observation

A::Matrix{<:Real}=initialize_transition_matrix(K): The transition matrix of the HMM (defaults to random initialization)

πₖ::Vector{Float64}=initialize_state_distribution(K): The initial state distribution of the HMM (defaults to random initialization)

Returns

::HiddenMarkovModel: Hidden Markov Model Object with Gaussian Emissions

```

StateSpaceDynamics.SwitchingBernoulliRegression — Method

SwitchingBernoulliRegression(; K::Int, input_dim::Int, include_intercept::Bool=true, β::Vector{<:Real}=if include_intercept zeros(input_dim + 1) else zeros(input_dim) end, λ::Float64=0.0, A::Matrix{<:Real}=initialize_transition_matrix(K), πₖ::Vector{Float64}=initialize_state_distribution(K))

Create a Switching Bernoulli Regression Model

Arguments

K::Int: The number of hidden states.

input_dim::Int: The dimensionality of the input data.

include_intercept::Bool=true: Whether to include an intercept in the regression model (defaults to true).

β::Vector{<:Real}: The regression coefficients (defaults to zeros).

λ::Float64=0.0: Regularization parameter for the regression (defaults to zero).

A::Matrix{<:Real}=initialize_transition_matrix(K): The transition matrix of the HMM (defaults to random initialization).

πₖ::Vector{Float64}=initialize_state_distribution(K): The initial state distribution of the HMM (defaults to random initialization).

Returns

::HiddenMarkovModel: A Switching Bernoulli Regression Model

StateSpaceDynamics.SwitchingGaussianRegression — Method

SwitchingGaussianRegression(; 
    K::Int,
    input_dim::Int,
    output_dim::Int,
    include_intercept::Bool = true,
    β::Matrix{<:Real} = if include_intercept
        zeros(input_dim + 1, output_dim)
    else
        zeros(input_dim, output_dim)
    end,
    Σ::Matrix{<:Real} = Matrix{Float64}(I, output_dim, output_dim),
    λ::Float64 = 0.0,
    A::Matrix{<:Real} = initialize_transition_matrix(K),
    πₖ::Vector{Float64} = initialize_state_distribution(K)
)

Create a Switching Gaussian Regression Model

Arguments

K::Int: The number of hidden states.

input_dim::Int: The dimensionality of the input features.

output_dim::Int: The dimensionality of the output predictions.

include_intercept::Bool: Whether to include an intercept in the regression model (default is true).

β::Matrix{<:Real}: The regression coefficients (defaults to zeros based on input_dim and output_dim).

Σ::Matrix{<:Real}: The covariance matrix of the Gaussian emissions (defaults to an identity matrix).

λ::Float64: The regularization parameter for the regression (default is 0.0).

A::Matrix{<:Real}: The transition matrix of the Hidden Markov Model (defaults to random initialization).

πₖ::Vector{Float64}: The initial state distribution of the Hidden Markov Model (defaults to random initialization).

Returns

::HiddenMarkovModel: A Switching Gaussian Regression Model