Misc

StateSpaceDynamics.ProbabilisticPCA — Type

mutable struct ProbabilisticPCA

Probabilistic PCA model from Bishop's Pattern Recognition and Machine Learning.

Fields:

W: Weight matrix that maps from latent space to data space.
σ²: Noise variance
μ: Mean of the data
k: Number of latent dimensions
D: Number of features
z: Latent variables

StateSpaceDynamics.ProbabilisticPCA — Method

ProbabilisticPCA(;W::Matrix{<:AbstractFloat}, σ²:: <: AbstractFloat, μ::Matrix{<:AbstractFloat}, k::Int, D::Int)

Constructor for ProbabilisticPCA model.

# Args:

- W::Matrix{<:AbstractFloat}: Weight matrix that maps from latent space to data space.

- σ²:: <: AbstractFloat: Noise variance

- μ::Matrix{<:AbstractFloat}: Mean of the data

- k::Int: Number of latent dimensions

- D::Int: Number of features

# Example:

```julia

# PPCA with unknown parameters

ppca = ProbabilisticPCA(k=1, D=2)

# PPCA with known parameters

ppca = ProbabilisticPCA(W=rand(2, 1), σ²=0.1, μ=rand(2), k=1, D=2)

```

StateSpaceDynamics.E_Step — Method

E_Step(ppca::ProbabilisticPCA, X::Matrix{<:AbstractFloat})

Expectation step of the EM algorithm for PPCA. See Bishop's Pattern Recognition and Machine Learning for more details.

Args:

ppca::ProbabilisticPCA: PPCA model

X::Matrix{<:AbstractFloat}: Data matrix

Examples:

ppca = ProbabilisticPCA(K=1, D=2)
E_Step(ppca, rand(10, 2))

StateSpaceDynamics.M_Step! — Method

M_Step!(model::ProbabilisticPCA, X::Matrix{<:AbstractFloat}, E_z::Matrix{<:AbstractFloat}, E_zz::Array{<:AbstractFloat, 3}

Maximization step of the EM algorithm for PPCA. See Bishop's Pattern Recognition and Machine Learning for more details.

Args:

model::ProbabilisticPCA: PPCA model

X::Matrix{<:AbstractFloat}: Data matrix

E_z::Matrix{<:AbstractFloat}: E[z]

E_zz::Matrix{<:AbstractFloat}: E[zz']

Examples:

ppca = ProbabilisticPCA(K=1, D=2)
E_z, E_zz = E_Step(ppca, rand(10, 2))
M_Step!(ppca, rand(10, 2), E_z, E_zzᵀ)

StateSpaceDynamics.fit! — Function

fit!(model::ProbabilisticPCA, X::Matrix{<:AbstractFloat}, max_iter::Int=100, tol::AbstractFloat=1e-6)

Fit the PPCA model to the data using the EM algorithm.

Args:

model::ProbabilisticPCA: PPCA model

X::Matrix{<:AbstractFloat}: Data matrix

max_iter::Int: Maximum number of iterations

tol::AbstractFloat: Tolerance for convergence

Examples:

ppca = ProbabilisticPCA(K=1, D=2)
fit!(ppca, rand(10, 2))

StateSpaceDynamics.loglikelihood — Method

loglikelihood(model::ProbabilisticPCA, X::Matrix{<:AbstractFloat})

Calculate the log-likelihood of the data given the PPCA model.

Args:

model::ProbabilisticPCA: PPCA model

X::Matrix{<:AbstractFloat}: Data matrix

Examples:

ppca = ProbabilisticPCA(K=1, D=2)
loglikelihood(ppca, rand(10, 2))

StateSpaceDynamics.block_tridgm — Method

block_tridgm(main_diag::Vector{Matrix{T}}, upper_diag::Vector{Matrix{T}}, lower_diag::Vector{Matrix{T}}) where {T<:Real}

Construct a block tridiagonal matrix from three vectors of matrices.

Arguments

main_diag::Vector{Matrix{T}}: Vector of matrices for the main diagonal.

upper_diag::Vector{Matrix{T}}: Vector of matrices for the upper diagonal.

lower_diag::Vector{Matrix{T}}: Vector of matrices for the lower diagonal.

Returns

A sparse matrix representing the block tridiagonal matrix.

Throws

ErrorException if the lengths of upper_diag and lower_diag are not one less than the length of main_diag.

StateSpaceDynamics.block_tridiagonal_inverse — Method

block_tridiagonal_inverse(A, B, C)

Compute the inverse of a block tridiagonal matrix.

Arguments

A: Lower diagonal blocks.

B: Main diagonal blocks.

C: Upper diagonal blocks.

Returns

λii: Diagonal blocks of the inverse.

λij: Off-diagonal blocks of the inverse.

Notes: This implementation is from the paper:

"An Accelerated Lambda Iteration Method for Multilevel Radiative Transfer” Rybicki, G.B., and Hummer, D.G., Astronomy and Astrophysics, 245, 171–181 (1991), Appendix B.

StateSpaceDynamics.block_tridiagonal_inverse_static — Method

block_tridiagonal_inverse_static(A, B, C)

Compute the inverse of a block tridiagonal matrix using static matrices. See block_tridiagonal_inverse for details.

StateSpaceDynamics.euclidean_distance — Method

euclidean_distance(a::AbstractVector{Float64}, b::AbstractVector{Float64})

Calculate the Euclidean distance between two points.

Arguments

a::AbstractVector{Float64}: The first point.

b::AbstractVector{Float64}: The second point.

Returns

The Euclidean distance between a and b.

StateSpaceDynamics.gaussian_entropy — Method

gaussian_entropy(H::Symmetric{T}) where T <: Real

Calculate the entropy of a Gaussian distribution with Hessian (i.e. negative precision) matrix H.

Arguments

H::Symmetric{T}: The Hessian matrix.

Returns

The entropy of the Gaussian distribution.

StateSpaceDynamics.kmeans_clustering — Function

kmeans_clustering(data::Vector{Float64}, k_means::Int, max_iters::Int=100, tol::Float64=1e-6)

Perform K-means clustering on vector data.

Arguments

data::Vector{Float64}: The input data vector.

k_means::Int: The number of clusters.

max_iters::Int=100: Maximum number of iterations.

tol::Float64=1e-6: Convergence tolerance.

Returns

A tuple containing the final centroids and cluster labels for each data point.

StateSpaceDynamics.kmeans_clustering — Function

kmeans_clustering(data::Matrix{<:Real}, k_means::Int, max_iters::Int=100, tol::Float64=1e-6)

Perform K-means clustering on the input data.

Arguments

data::Matrix{<:Real}: The input data matrix where each row is a data point.

k_means::Int: The number of clusters.

max_iters::Int=100: Maximum number of iterations.

tol::Float64=1e-6: Convergence tolerance.

Returns

A tuple containing the final centroids and cluster labels for each data point.

StateSpaceDynamics.kmeanspp_initialization — Method

kmeanspp_initialization(data::Matrix{<:Real}, k_means::Int)

Perform K-means++ initialization for cluster centroids.

Arguments

data::Matrix{<:Real}: The input data matrix where each row is a data point.

k_means::Int: The number of clusters.

Returns

A matrix of initial centroids for K-means clustering.

StateSpaceDynamics.kmeanspp_initialization — Method

kmeanspp_initialization(data::Vector{Float64}, k_means::Int)

Perform K-means++ initialization for cluster centroids on vector data.

Arguments

data::Vector{Float64}: The input data vector.

k_means::Int: The number of clusters.

Returns

A matrix of initial centroids for K-means clustering.

StateSpaceDynamics.logistic — Method

logistic(x::Real)

Calculate the logistic function in a numerically stable way.

Arguments

x::Real: The input value.

Returns

The result of the logistic function applied to x.

StateSpaceDynamics.make_posdef! — Method

make_posdef!(A::Matrix{T}) where {T}

Ensure that a matrix is positive definite by adjusting its eigenvalues.

Arguments

A::Matrix{T}: The input matrix.

Returns

A positive definite matrix derived from A.

StateSpaceDynamics.row_matrix — Method

row_matrix(x::AbstractVector)

Convert a vector to a row matrix.

Arguments

x::AbstractVector: The input vector.

Returns

A row matrix (1 × n) containing the elements of x.

StateSpaceDynamics.stabilize_covariance_matrix — Method

stabilize_covariance_matrix(Σ::Matrix{<:Real})

Stabilize a covariance matrix by ensuring it is symmetric and positive definite.

Arguments

Σ::Matrix{<:Real}: The input covariance matrix.

Returns

A stabilized version of the input covariance matrix.