Emission Models

The StateSpaceDynamics.jl package provides several emission models for state space modeling. These models define how observations are generated from latent states.

StateSpaceDynamics.AutoRegressionEmission — Type

AutoRegressionEmission <: EmissionModel

Store an autoregressive emission model, which wraps around a GaussianRegressionEmission.

Fields

output_dim::Int: The dimensionality of the output data

order::Int: The order of the Autoregressive process

innerGaussianRegression::GaussianRegressionEmission: The underlying Gaussian regression model used for the emissions.

StateSpaceDynamics.AutoRegressionEmission — Method

AutoRegressionEmission(output_dim, order, include_intercept, β, Σ, λ)

Create an Autoregressive emission model.

Arguments

- `output_dim::Int`: The dimensionality of the output data.
- `order::Int`: The order of the autoregressive process.
- `include_intercept::Bool`: Whether to include a regression intercept.
- `β::AbstractMatrix`: The regression coefficient matrix.
- `Σ::AbstractMatrix`: The regression covariance matrix.
- `λ<:Real`: The L2 regularization parameter.

Returns

- `model::AutoRegressionEmission`: The autoregressive emission model.

StateSpaceDynamics.BernoulliRegressionEmission — Type

BernoulliRegressionEmission

Store a Bernoulli regression model.

Fields

input_dim::Int: Dimensionality of the input data.

output_dim::Int: Dimensionality of the outputd data.

include_intercept::Bool: Whether to include an intercept term.

β::AbstractMatrix{<:Real}: Bernoulli regression coefficients.

λ<:Real: L2 Regularization parameter.

```

StateSpaceDynamics.BernoulliRegressionEmission — Method

BernoulliRegressionEmission(Args)

Create a Bernoulli regression emission model.

Arguments

- `input_dim::Int`: Dimensionality of the input dimension
- `output_dim::Int`: Dimensionality of the output dimension
- `include_intercept::Bool`: Whether to include a regression intercept.
- `β::AbstractMatrix`: The regression coefficient matrix.
- `Σ::AbstractMatrix`: The regression covariance matrix.
- `λ<:Real`: The L2 regularization parameter.

Returns

- `model::BernoulliRegressionEmission`: The Bernoulli regression emission model.

StateSpaceDynamics.GaussianEmission — Type

mutable struct GaussianEmission <: EmissionModel

GaussianEmission model with mean and covariance.

StateSpaceDynamics.GaussianEmission — Method

function GaussianEmission(; output_dim::Int, μ::AbstractVector, Σ::AbstractMatrix)

Create a GaussianEmission model.

Arguments

output_dim::Int: The output dimension of the emission

μ::AbstractVector: The mean of the Gaussian

Σ::AbstractMatrix: The covariance matrix of the Gaussian

Returns

GaussianEmission<:EmissionModel: The Gaussian emission model

```

StateSpaceDynamics.GaussianRegressionEmission — Type

GaussianRegressionEmission

Store a Gaussian regression Emission model.

Fields

input_dim::Int: Dimension of the input data.

output_dim::Int: Dimension of the output data.

include_intercept::Bool: Whether to include an intercept term; if true, the first column of β is assumed to be the intercept/bias.

β::AbstractMatrix{<:Real} = if include_intercept zeros(input_dim + 1, output_dim) else zeros(input_dim, output_dim) end: Coefficient matrix of the model. Shape inputdim by outputdim. The first row are the intercept terms, if included.

Σ::AbstractMatrix{<:Real}: Covariance matrix of the model.

λ:<Real: Regularization parameter.

StateSpaceDynamics.GaussianRegressionEmission — Method

GaussianRegressionEmission(input_dim, output_dim, include_intercept, β, Σ, λ)

Create a Gaussian regression emission model.

Arguments

-`input_dim::Int`: Dimensionality of input data
-`output_dim::Int`: Dimensionality of output data
-`include_intercept:Bool`: Whether to include a regression intercept
-`β::AbstractMatrix`: The regression coefficient matrix
-`Σ::AbstractMatrix`: The regression covariance matrix
-`λ<:Real`: The L2 regularization parameter

Returns

-`model::GaussianRegressionEmission`: The Gaussian regression emission model

StateSpaceDynamics.PoissonRegressionEmission — Type

PoissonRegressionEmission

A Poisson regression model.

Fields

input_dim::Int: Dimensionality of the input data.

output_dim::Int: Dimensionality of the output data.

include_intercept::Bool: Whether to include a regression intercept.

β::AbstractMatrix{<:Real}: The regression coefficients matrix.

λ::Real;: L2 Regularization parameter.

StateSpaceDynamics.PoissonRegressionEmission — Method

PoissonRegressionEmission(Args)

Create a Poisson regression emission model.

Arguments

- `input_dim::Int`: Dimensionality of the input data.
- `output_dim::Int`: Dimensionality of the output data.
- `include_intercept::Bool`: Whether to include a regression intercept.
- `β::AbstractMatrix{<:Real}`: The regression coefficients matrix.
- `λ::Real;`: L2 Regularization parameter.

Returns

- `model::PoissonRegressionEmission`: The Poisson regression emission model.

StateSpaceDynamics.RegressionOptimization — Type

RegressionOptimization{T<:RegressionEmission}

Hold the optimization problem data for regression emissions.

Base.rand — Method

Random.rand(rng::AbstractRNG, model::AutoRegressionEmission, X::Matrix{<:Real})

Generate samples from an autoregressive emission model.

Arguments

rng::AbstractRNG: The seed.

model::AutoRegressionEmission: The autoregressive emission model.

X::Matrix{<:Real}: The current data from which to generate samples.

Returns

Matrix{Float64}: The updated observation sequence with the new sample appended.

Base.rand — Method

Random.rand(rng::AbstractRNG, model::BernoulliRegressionEmission, Φ::Union{Matrix{<:Real},Vector{<:Real}})

Generate samples from a Bernoulli regression model.

Arguments

- `rng::AbstractRNG`: The seed.
- `model::BernoulliRegressionEmission`: Bernoulli regression model.
- `Φ::AbstractMatrix{<:Real}`: Design matrix of shape `(n, input_dim)`.

Returns

Y::AbstractMatrix{<:Real}: Matrix of samples of shape (n, output_dim).

Base.rand — Method

 Random.rand(rng::AbstractRNG, model::GaussianEmission; n::Int=1)

Generate n samples from a Gaussian emission model.

Arguments

- `rng::AbstractRNG`: Seed
-`model::GaussianEmission`: The GaussianEmission model
-`n::Int=1`: The number of samples to generate

Returns

-`samples::Matrix{<:Real}`: Matrix of samples (n, output_dim)

Base.rand — Method

 Random.rand(rng::AbstractRNG, model::GaussianRegressionEmission, Φ::Union{Matrix{<:Real},Vector{<:Real}})

Generate samples from a Gaussian regression emission model.

Arguments

- `rng::AbstractRNG`: Seed
-`model::GaussianRegressionEmission`: The Gaussian regression model
-`Φ::Union{Matrix{<:Real},Vector{<:Real}}`: The input data (defines number of samples to generate)

Returns

-`samples::Matrix{<:Real}`: Matrix of samples (n, output_dim)

Base.rand — Method

Random.rand(rng::AbstractRNG, model::PoissonRegressionEmission, Φ::Union{Matrix{<:Real},Vector{<:Real}})

Generate samples from a Poisson regression model.

Arguments

- `rng::AbstractRNG`: The seed.
- `model::PoissonRegressionEmission`: Poisson regression model.
- `Φ::AbstractMatrix{<:Real}`: Design matrix of shape `(n, input_dim)`.

Returns

Y::AbstractMatrix{<:Real}: Matrix of samples of shape (n, output_dim).

Base.rand — Method

Random.rand(model::AutoRegressionEmission, X::Matrix{<:Real})

Generate samples from an autoregressive emission model.

Base.rand — Method

Random.rand(model::BernoulliRegressionEmission, Φ::Union{Matrix{<:Real},Vector{<:Real}})

Generate samples from a Bernoulli regression emission.

Base.rand — Method

Random.rand(model::GaussianEmission; kwargs...)

Generate random samples from a Gaussian emission model.

Base.rand — Method

Random.rand(model::GaussianRegressionEmission, Φ::Union{Matrix{<:Real},Vector{<:Real}})

Generate samples from a Gaussian regression model.

Base.rand — Method

Random.rand(model::PoissonRegressionEmission, Φ::Union{Matrix{<:Real},Vector{<:Real}})

Generate samples from a Poisson regression emission model.

StateSpaceDynamics.calc_regularization — Method

calc_regularization(β::AbstractMatrix{T1}, λ::T2, include_intercept::Bool=true) where {T1<:Real, T2<:Real}

Calculate L2 regularization term for regression coefficients.

StateSpaceDynamics.calc_regularization_gradient — Method

function calc_regularization_gradient(β::AbstractMatrix{T1}, λ::T2, include_intercept::Bool=true) where {T1<:Real, T2<:Real}

Calculate gradient of L2 regularization term for regression coefficients.

StateSpaceDynamics.construct_AR_feature_matrix — Method

construct_AR_feature_matrix(data::AbstractMatrix{T}, order::Int, include_intercept=false) where {T<:Real}

Construct an autoregressive (AR) feature matrix from input time series data.

Arguments

data::AbstractMatrix{<:Real}: A matrix of size (num_feats, T), where num_feats is the number of features, and T is the number of timepoints.

order::Int: The autoregressive order, determining how many past timepoints are included for each time step.

include_intercept::Bool=false: Whether to include an intercept regression term.

Returns

Matrix{Float64}: A transformed feature matrix of size (num_feats * (order + 1), T - order), where each column contains stacked feature vectors from the current and past order timepoints.

StateSpaceDynamics.create_optimization — Function

create_optimization(
    model::RegressionEmission,
    X::AbstractMatrix{<:Real},
    y::AbstractMatrix{<:Real},
    w::V)

Create regression optimization problem.

StateSpaceDynamics.fit! — Method

fit!(
    model::RegressionEmission,
    X::AbstractMatrix{T1},
    y::AbstractMatrix{T2},
    w::AbstractVector{T3}=ones(size(y, 1))) where {T1<:Real, T2<:Real, T3<:Real}

Fit a regression emission model give input data X, output data y, and weights w.

Arguments

- `model::RegressionEmission`: A regression emission model.
- `X::AbstractMatrix{<:Real}:`: Input data.
- `y::AbstractMatrix{<:Real}`: Output data.
- `w::AbstractVector{<:Real}`: Weights to define each point's contribution to the fit.

Returns

- `model::RegressionEmission`: The regression model with the newly updated parameters.

StateSpaceDynamics.fit! — Method

function fit!(model::GaussianEmission, 
        Y::AbstractMatrix{T}, 
        w::AbstractVector{T}=ones(size(Y, 1))) where {T<:Real}

Fit a GaussianEmission model to the data Y weighted by weights w.

StateSpaceDynamics.loglikelihood — Function

function loglikelihood(
    model::BernoulliRegressionEmission,
    Φ::AbstractMatrix{T1},
    Y::AbstractMatrix{T2},
    w::AbstractVector{T3}=ones(size(Y, 1))) where {T1<:Real, T2<:Real, T3<:Real}

Calculate the log likelihood of the data Y given the Bernoulli regression emission model and the input features Φ. Optionally, a vector of weights w can be provided.

StateSpaceDynamics.loglikelihood — Function

loglikelihood(
    model::PoissonRegressionEmission,
    Φ::AbstractMatrix{T1},
    Y::AbstractMatrix{T2},
    w::AbstractVector{T3}=ones(size(Y, 1))) where {T1<:Real, T2<:Real, T3<:Real}

Calculate the log-likelihood of a Poisson regression model.

StateSpaceDynamics.loglikelihood — Method

loglikelihood(
    model::AutoRegressionEmission,
    X::AbstractMatrix{T},
    Y::AbstractMatrix{T},
    w::Vector{T}=ones(size(Y, 1))) where {T<:Real}

Calculate the log likelihood of the data Y given the autoregressive emission model and the previous observations X.

StateSpaceDynamics.loglikelihood — Method

loglikelihood(model::GaussianEmission, Y::AbstractMatrix{T}) where {T<:Real}

Calculate the log likelihood of the data Y given the Gaussian emission model.

StateSpaceDynamics.loglikelihood — Method

loglikelihood(model::GaussianRegressionEmission,
    Φ::AbstractMatrix{T},
    Y::AbstractMatrix{T},
    w::AbstractVector{T}=ones(size(Y, 1))) where {T<:Real}

Calculate the log likelihood of the data Y given the Gaussian regression emission model and the input features Φ.

StateSpaceDynamics.objective — Method

objective(
    opt::RegressionOptimization{<:BernoulliRegressionEmission},
    β_vec::Vector{T}) where {T<:Real}

Define the objective function for a Bernoulli regression emission model.

StateSpaceDynamics.objective — Method

objective(
opt::RegressionOptimization{<:PoissonRegressionEmission},
β_vec::AbstractVector{T}) where {T<:Real}

Define the objective function for a Poisson regression emission.

StateSpaceDynamics.objective — Method

objective(
    opt::Union{RegressionOptimization{<:GaussianRegressionEmission}, 
    RegressionOptimization{<:AutoRegressionEmission}}, 
    β_vec::AbstractVector{T}) where {T<:Real}

Define the objective function for Gaussian/AR regression emission models.

StateSpaceDynamics.objective_gradient! — Method

objective_gradient!(
    G::AbstractVector{T},
    opt::RegressionOptimization{<:BernoulliRegressionEmission},
    β_vec::AbstractVector{T}) where {T<:Real}

Define the objective gradient for a Bernoulli regression emission model.

StateSpaceDynamics.objective_gradient! — Method

objective_gradient!(
    G::AbstractVector{T},
    opt::RegressionOptimization{<:PoissonRegressionEmission},
    β_vec::AbstractVector{T}) where {T<:Real}

Define the objective gradient for a Poisson regression emission.

StateSpaceDynamics.objective_gradient! — Method

objective_gradient!(
    G::AbstractVector{T},
    opt::Union{RegressionOptimization{<:GaussianRegressionEmission}, RegressionOptimization{<:AutoRegressionEmission}},
    β_vec::AbstractVector{T}) where {T<:Real}

Define the gradient of the objective function for Gaussian/AR regression emission models

StateSpaceDynamics.post_optimization! — Method

post_optimization!(model::GaussianRegressionEmission, opt::RegressionOptimization)

Stabilize the covariance matrix for GaussianRegressionEmissions.